+++ PolitiekeBarometer_MVC/Views/Home/Index.cshtml
@@ -24,7 +24,7 @@
         <img src="http://users.telenet.be/excelleren/images/Grafieken/Grafieken1.jpg" alt="Grafiek1" class="graph" />
         <a href="#"><img src="https://cdn4.iconfinder.com/data/icons/wirecons-free-vector-icons/32/menu-alt-256.png" alt="Menu-Icon" class="icon" /></a>
     </div>
-    <div class="inline">
+    <div class="inline"> 
         <img src="https://informaticalessen.be/wp-content/uploads/2017/01/Oefening-Excel-of-Spreadsheets-Lijndiagram-of-lijngrafiek.png" alt="Grafiek2" class="graph" />
         <a href="#"><img src="https://cdn4.iconfinder.com/data/icons/wirecons-free-vector-icons/32/menu-alt-256.png" alt="Menu-Icon" class="icon" /></a>
     </div>
