+++ PolitiekeBarometer_MVC/Views/Home/_DashboardPartial.cshtml
@@ -7,8 +7,8 @@
 <div id="dashboard">
     @if (Request.IsAuthenticated)
     {
-      @Html.ActionLink("","DashboardController","getDashboard")
-        <ul id="tabs" runat="server">
+        <ul id="tabs">
+        
         @foreach (var item in Model)
          {
             <li>@Html.DisplayFor(modelItem => item.Naam)</li>
@@ -16,5 +16,10 @@
             <li class="inline"><a href="#"><img src="http://icons.iconarchive.com/icons/icons8/android/256/Very-Basic-Plus-icon.png" alt="Plus-Icon" class="icon" /></a></li>
         </ul>
     }
-    
+    else
+    {
+        <div>Standaard</div>
+      
+    }
+
 </div>
